+++ text-messaging/getTextReminders.js
@@ -10,7 +10,6 @@ function loadTextReminders(){
   try
   {
     // Load client secrets from a local file.
-    // var content = fs.readFileSync(CREDENTIALS, 'utf8')
     return authorize(JSON.parse(CREDENTIALS), main);
   }
   catch(err)
@@ -27,46 +26,18 @@ function authorize(credentials, callback) {
 
   // Check if we have previously stored a token.
   try{
-    // var token = fs.readFileSync(TOKEN, 'utf8');
     oAuth2Client.setCredentials(JSON.parse(TOKEN));
     return callback(oAuth2Client);
 }
   catch(err)
   {
     console.log("Here's an error: "+err);
-      // return getNewToken(oAuth2Client, callback);
   }
 }
 
-
-// function getNewToken(oAuth2Client, callback) {
-//   const authUrl = oAuth2Client.generateAuthUrl({
-//     access_type: 'offline',
-//     scope: SCOPES,
-//   });
-//   console.log('Authorize this app by visiting this url:', authUrl);
-//   const rl = readline.createInterface({
-//     input: process.stdin,
-//     output: process.stdout,
-//   });
-//   rl.question('Enter the code from that page here: ', (code) => {
-//     rl.close();
-//     oAuth2Client.getToken(code, (err, token) => {
-//       if (err) return console.error('Error while trying to retrieve access token', err);
-//       oAuth2Client.setCredentials(token);
-//       // Store the token to disk for later program executions
-//       fs.writeFile(TOKEN_PATH, JSON.stringify(token), (err) => {
-//         if (err) return console.error(err);
-//         console.log('Token stored to', TOKEN_PATH);
-//       });
-//       callback(oAuth2Client);
-//     });
-//   });
-// }
-
-
 async function getMessages(auth) {
 
+  //get messages from Google Sheets
   const sheets = google.sheets({ version: "v4", auth });
   const daily_check = await sheets.spreadsheets.values.get({
     spreadsheetId: process.env.SPREADSHEET_ID,
